+++ src/Typography/Link/Link.stories.tsx
@@ -2,12 +2,18 @@ import * as React from 'react';
 import { storiesOf } from '@storybook/react';
 import { action } from '@storybook/addon-actions';
 import { Link } from './Link.component';
+import { boolean } from '@storybook/addon-knobs/react';
 
 storiesOf('Typography/Link', module)
   .add(
     'Standalone',
     () => (
-      <Link solo={true} href="https://www.tkxs.com" target="_blank">
+      <Link
+        solo={true}
+        href="https://www.tkxs.com"
+        target="_blank"
+        disabled={boolean('disabled', false)}
+      >
         Read More
       </Link>
     ),